# docs/swagger.yaml
basePath: /
definitions:
  queries.AuthorizedUserView:
    properties:
      company_id:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        type: string
    type: object
  request.CreateReservationRequest:
    properties:
      couponCode:
        type: string
      endTime:
        type: string
      note:
        type: string
      resourceId:
        type: string
      startTime:
        type: string
    required:
    - endTime
    - resourceId
    - startTime
    type: object
...
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  response.LoginResponse:
    properties:
      user:
        $ref: '#/definitions/queries.AuthorizedUserView'
    type: object
  response.ReservationListResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      priceCents:
        type: integer
      resourceId:
        type: string
      resourceName:
        type: string
      slot:
        type: string
      status:
        type: string
    type: object
  response.ReservationResponse:
    properties:
      couponCode:
        type: string
      couponId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      note:
        type: string
      priceCents:
        type: integer
      resourceId:
        type: string
      resourceName:
        type: string
      slot:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      userEmail:
        type: string
      userId:
        type: string
    type: object
...
info:
  contact: {}
  description: JWT Authorization header using the Bearer scheme
  title: Gin Clean Starter
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout current user session
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/me:
    get:
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queries.AuthorizedUserView'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refresh access token using refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
...
  /reservations:
    get:
      description: Get all reservations for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ReservationListResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Create a new reservation with idempotency key
      parameters:
      - description: Idempotency key for duplicate prevention
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Reservation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservations
  /reservations/{id}:
    get:
      description: Get reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get reservation
      tags:
      - reservations
...
schemes:
- http
- https
swagger: "2.0"